requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
django-filter==2.4.0
djangorestframework==3.12.4
idna==2.10
importlib-metadata==4.0.1
Markdown==3.3.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.4
zipp==3.4.1
